+++ lib/shared/src/chat/preamble.ts
@@ -1,5 +1,11 @@
 import { Message } from '../sourcegraph-api'
 
+export interface Preamble {
+    actions: string
+    rules: string
+    answer: string
+}
+
 const actions = `You are Cody, an AI-powered coding assistant created by Sourcegraph. You work inside a text editor. You have access to my currently open files. You perform the following actions:
 - Answer general programming questions.
 - Answer questions about the code that I have provided to you.
@@ -31,9 +37,12 @@ I will acknowledge when I don't know an answer or need more context.`
  * Creates and returns an array of two messages: one from a human, and the supposed response from the AI assistant.
  * Both messages contain an optional note about the current codebase if it's not null.
  */
-export function getPreamble(codebase: string | undefined): Message[] {
-    const preamble = [actions, rules]
-    const preambleResponse = [answer]
+export function getPreamble(codebase: string | undefined, customPreamble?: Preamble): Message[] {
+    const actionsText = customPreamble?.actions ?? actions
+    const rulesText = customPreamble?.rules ?? rules
+    const answerText = customPreamble?.answer ?? answer
+    const preamble = [actionsText, rulesText]
+    const preambleResponse = [answerText]
 
     if (codebase) {
         const codebasePreamble =
@@ -59,9 +68,12 @@ export function getPreamble(codebase: string | undefined): Message[] {
     ]
 }
 
-export function getMultiRepoPreamble(codebases: string[]): Message[] {
-    const preamble = [actions, multiRepoRules]
-    const preambleResponse = [answer]
+export function getMultiRepoPreamble(codebases: string[], customPreamble?: Preamble): Message[] {
+    const actionsText = customPreamble?.actions ?? actions
+    const rulesText = customPreamble?.rules ?? multiRepoRules
+    const answerText = customPreamble?.answer ?? answer
+    const preamble = [actionsText, rulesText]
+    const preambleResponse = [answerText]
 
     if (codebases.length === 1) {
         return getPreamble(codebases[0])